Renamed everything to be in-line with gem name A gem needs to be consistent If the gem is called ruby-nessus, require 'ruby-nessus' needs to work which it now will Reason for change is another gem also uses require 'nessus' They also define a module Nessus so this module has also been fixed to module RubyNessus
