--- lib/nessus.rb
@@ -1,3 +1,0 @@
-# Ruby-Nessus Files
-require 'nessus/nessus'
-require 'nessus/version'
