--- lib/nessus/nessus.rb
@@ -1,6 +1,0 @@
-require 'nessus/parse'
-require 'nessus/core_ext/helpers'
-
-module Nessus
-
-end
