+++ lib/ruby-nessus.rb
@@ -1,0 +1,3 @@
+# Ruby-Nessus Files
+require 'ruby-nessus/ruby-nessus'
+require 'ruby-nessus/version'
