+++ lib/ruby-nessus/Version1/port.rb
@@ -1,4 +1,4 @@
-module Nessus
+module RubyNessus
   module Version1
     
     class Port
