+++ lib/ruby-nessus/core_ext/helpers.rb
