+++ lib/ruby-nessus/ruby-nessus.rb
@@ -1,0 +1,6 @@
+require 'ruby-nessus/parse'
+require 'ruby-nessus/core_ext/helpers'
+
+module RubyNessus
+
+end
