+++ lib/ruby-nessus/version.rb
@@ -1,3 +1,3 @@
-module Nessus
+module RubyNessus
   VERSION = '1.2.0'
 end
