+++ spec/Version2/event_spec.rb
@@ -4,7 +4,7 @@ require 'helpers/xml'
 describe "Nessus Version 2: Event" do
   
   before(:all) do
-    @xml = Nessus::Version2::XML.new(Helpers::DOT_NESSUS_V2)
+    @xml = RubyNessus::Version2::XML.new(Helpers::DOT_NESSUS_V2)
     @host = @xml.hosts.first
     @bad_event = @host.events.first
     @good_event = @host.events.last
