+++ spec/ruby-nessus_spec.rb
@@ -1,9 +1,9 @@
-require 'nessus/version'
+require 'ruby-nessus/version'
  
 require 'spec_helper'
  
-describe Nessus do
+describe RubyNessus do
   it "should define a VERSION constant" do
-    Nessus.should be_const_defined('VERSION')
+    RubyNessus.should be_const_defined('VERSION')
   end
 end
