+++ spec/spec_helper.rb
@@ -1,4 +1,4 @@
 require 'rubygems'
 gem 'rspec', '>=1.1.12'
  
-require "nessus"
+require "ruby-nessus"
